Completed Version
This is what I would like to turn in for my final project, if you are okay with it. It is a completed screen, and it has the basics of all things I was hoping to add.

I do have some follow-up questions, in hopes of continuing to learn to code.

1) How do I get the header to stop scrolling?
2)How do I get the SVGs to stop clipping when I don't have "overflow:visible" on them?
3) I'm happy with the amount of space below the h2s for now, but I was wondering how I might get it to be less if need be.
4) How do I get the second image to crop and not squish?